Add clang-tidy to pre-commit hooks and CI
- Add .clang-tidy configuration with enabled checks for bugprone,
  clang-analyzer, misc, modernize, performance, and readability
- Add clang-tidy pre-commit hook with fallback to brew's llvm
- Add separate lint-cpp CI job that runs in parallel with build-and-test
- Update CMakeLists.txt to export compile_commands.json and set sysroot
- Fix code issues found by clang-tidy:
  - Fix narrowing conversions with explicit static_cast
  - Remove unused variables (two, neg_one, edgePaddingHigh, sliceSizes)
  - Fix nullptr usage
  - Fix container size checks (.size() >= 1 → !.empty())
  - Simplify duplicate branch logic
- Exclude pjrt_executable.cc from clang-tidy (requires generated proto)
- Add .cache/ to .gitignore

+++ src/pjrt_plugin/ops/bitwise_ops.mm
@@ -60,7 +60,7 @@ static int getBitWidth(mlir::Operation* op) {
         return 0;
     auto elemType = tensorType.getElementType();
     if (auto intType = mlir::dyn_cast<mlir::IntegerType>(elemType)) {
-        return intType.getWidth();
+        return static_cast<int>(intType.getWidth());
     }
     return 0;
 }
